+++ ScenarioGUI/gui_classes/gui_structure_classes/button_box.py
@@ -7,6 +7,7 @@ from functools import partial as ft_partial
 from typing import TYPE_CHECKING
 
 import PySide6.QtWidgets as QtW  # type: ignore
+import PySide6.QtCore as QtC  # type: ignore
 
 import ScenarioGUI.global_settings as globs
 
@@ -70,7 +71,7 @@ class ButtonBox(Option):
                     ),
                 )
             )
-            button.clicked.connect(ft_partial(check, self.linked_options, self, self.get_value()))
+            button.toggled.connect(ft_partial(check, self.linked_options, self, self.get_value()))
 
     def get_value(self) -> int:
         """
@@ -165,7 +166,7 @@ class ButtonBox(Option):
         None
         """
         for button in self.widget:
-            button.clicked.connect(function_to_be_called)  # pylint: disable=E1101
+            button.toggled.connect(function_to_be_called)  # pylint: disable=E1101
 
     def set_text(self, name: str) -> None:
         """
@@ -201,6 +202,47 @@ class ButtonBox(Option):
             True if the linked "option" should be shown
         """
         return self.get_value() == value
+
+    def disable_entry(self, idx: int):
+        """
+        Disables the entry at index
+
+        Parameters
+        ----------
+        idx: int
+            index of entry which should be disabled
+        """
+        if self.widget[idx].isChecked():
+            self.widget[idx].setChecked(False)
+            self.widget[idx].setEnabled(False)
+            if self.widget[self.default_value].isEnabled():
+                self.widget[self.default_value].setChecked(True)
+            else:
+                widgets = [widget for widget in self.widget if widget.isEnabled()]
+                if widgets:
+                    widgets[0].setChecked(True)
+
+        self.widget[idx].setEnabled(False)
+        self.widget[idx].hide()
+        if not [widget for widget in self.widget if widget.isEnabled()]:
+            self.hide()
+        #if len([widget for widget in self.widget if widget.isEnabled()]) == 1:
+        #    [widget for widget in self.widget if widget.isEnabled()][0].setChecked(True)
+
+    def enable_entry(self, idx: int):
+        """
+        Enables the entry at index
+
+        Parameters
+        ----------
+        idx: int
+            index of entry which should be disabled
+        """
+        self.show()
+        self.widget[idx].setEnabled(True)
+        self.widget[idx].show()
+        if len([widget for widget in self.widget if widget.isEnabled()]) == 1:
+            self.widget[idx].setChecked(True)
 
     def create_widget(
         self,